[Ayla] Refactor power cycle reset
New features:

Changes:
- Refactory power cycle reset

Fixes:

Issues:
